Retrofit added for weather API

+++ Client-Apps/MathBasicGame/app/src/main/java/com/usk/mathbasicgame/model/Data.kt
@@ -1,5 +1,5 @@
 package com.usk.mathbasicgame.model
 
 class Data {
-    var temp : Int = 0
+    var temp : Double = 0.0
 }
